fix: map focus on line and stop selected
This one was tricky to pinpoint.
The algorithm for StopAndLineSelected.selectedStops() to focus the map on the selected stop plus previous and next was flawed.

We were removing the selected stop from the line stops in OnStopAndLineSelected, causing the algorithm to never find the stop. This case was not being handled, so we would mistakenly create a list with the first stop in the line (index -1 +1) and the selected stop.

The code in OnStopAndLineSelected that filters the selected stop was there to avoid showing a duplicated marker in the map, but had this unwanted side effect.

To avoid it, I removed the filter from the code so now lineStops contains all the stops. And we manually exclude the repeated stop in MarkerLayersByState.